Generate serialization of CFArrayRef https://bugs.webkit.org/show_bug.cgi?id=268700
Reviewed by Brady Eidson.

To retain existing behavior, we need a type distinct from CoreIPCArray which wraps NSArray.
The day may come when all our array serialization specifies what is inside the array, but
until then we need CoreIPCCFArray.

Covered by tests from 274048@main

* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/Shared/Cocoa/ArgumentCodersCocoa.h:
(IPC::ArgumentCoder<CoreIPCRetainPtr<T>>::encode):
(IPC::CoreIPCRetainPtr::CoreIPCRetainPtr): Deleted.
* Source/WebKit/Shared/Cocoa/ArgumentCodersCocoa.mm:
(IPC::encodeObjectDirectly<CFTypeRef>):
* Source/WebKit/Shared/Cocoa/CoreIPCCFType.h:
* Source/WebKit/Shared/Cocoa/CoreIPCCFType.serialization.in:
* Source/WebKit/Shared/Cocoa/CoreIPCRetainPtr.h: Copied from Source/WebKit/Shared/Cocoa/CoreIPCCFType.h.
(IPC::CoreIPCRetainPtr::CoreIPCRetainPtr):
* Source/WebKit/Shared/cf/ArgumentCodersCF.cpp:
(IPC::typeFromCFTypeRef):
(IPC::ArgumentCoder<CFArrayRef>::encode): Deleted.
(IPC::ArgumentCoder<RetainPtr<CFArrayRef>>::decode): Deleted.
* Source/WebKit/Shared/cf/ArgumentCodersCF.h:
* Source/WebKit/Shared/cf/CFTypes.serialization.in:
* Source/WebKit/Shared/cf/CoreIPCCFArray.h: Copied from Source/WebKit/Shared/Cocoa/CoreIPCCFType.h.
(WebKit::CoreIPCCFArray::CoreIPCCFArray):
(WebKit::CoreIPCCFArray::array const):
* Source/WebKit/Shared/cf/CoreIPCCFArray.mm: Copied from Source/WebKit/Shared/Cocoa/CoreIPCCFType.h.
(WebKit::CoreIPCCFArray::CoreIPCCFArray):
(WebKit::CoreIPCCFArray::createCFArray const):
* Source/WebKit/Shared/cf/CoreIPCCFArray.serialization.in: Copied from Source/WebKit/Shared/Cocoa/CoreIPCCFType.serialization.in.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/274062@main
