+++ Source/WebKit/Shared/cf/CoreIPCCFArray.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 Apple Inc. All rights reserved.
+ * Copyright (C) 2024 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -25,34 +25,27 @@
 
 #pragma once
 
-#if PLATFORM(COCOA)
+#if USE(CF)
 
-#include "CoreIPCRetainPtr.h"
-#include <wtf/ArgumentCoder.h>
+#include <wtf/RetainPtr.h>
+#include <wtf/UniqueRef.h>
+#include <wtf/Vector.h>
 
 namespace WebKit {
 
-class CoreIPCCFType {
-    WTF_MAKE_FAST_ALLOCATED;
+class CoreIPCCFType;
+
+class CoreIPCCFArray {
 public:
-    CoreIPCCFType(CFTypeRef cfType)
-        : m_cfType(cfType)
-    {
-    }
-
-    CoreIPCCFType(RetainPtr<CFTypeRef>&& cfType)
-        : m_cfType(WTFMove(cfType))
-    {
-    }
-
-    RetainPtr<id> toID() const { return (__bridge id)(m_cfType.get()); }
-
+    CoreIPCCFArray(CFArrayRef);
+    CoreIPCCFArray(Vector<UniqueRef<CoreIPCCFType>>&& array)
+        : m_array(WTFMove(array)) { }
+    RetainPtr<CFArrayRef> createCFArray() const;
+    const Vector<UniqueRef<CoreIPCCFType>>& array() const { return m_array; }
 private:
-    friend struct IPC::ArgumentCoder<CoreIPCCFType, void>;
-
-    IPC::CoreIPCRetainPtr<CFTypeRef> m_cfType;
+    Vector<UniqueRef<CoreIPCCFType>> m_array;
 };
 
 } // namespace WebKit
 
-#endif // PLATFORM(COCOA)
+#endif // USE(CF)
